chore(pii): improves logging for processor creation.
